Pedometer added
